+++ kitty/launcher/main.c
@@ -352,13 +352,6 @@ exec_kitten(int argc, char *argv[], char *exe_dir) {
     exit(1);
 }
 
-static void
-delegate_to_kitten_if_possible(int argc, char *argv[], char* exe_dir) {
-    if (argc > 1 && argv[1][0] == '@') exec_kitten(argc, argv, exe_dir);
-    if (argc > 2 && strcmp(argv[1], "+kitten") == 0 && is_wrapped_kitten(argv[2])) exec_kitten(argc - 1, argv + 1, exe_dir);
-    if (argc > 3 && strcmp(argv[1], "+") == 0 && strcmp(argv[2], "kitten") == 0 && is_wrapped_kitten(argv[3])) exec_kitten(argc - 2, argv + 2, exe_dir);
-}
-
 static bool
 is_boolean_flag(const char *x) {
     static const char *all_boolean_options = KITTY_CLI_BOOL_OPTIONS;
@@ -368,7 +361,7 @@ is_boolean_flag(const char *x) {
 }
 
 static void
-handle_fast_commandline(int argc, char *argv[]) {
+handle_fast_commandline(int argc, char *argv[], const char *instance_group_prefix) {
     char current_option_expecting_argument[128] = {0};
     CLIOptions opts = {0};
     int first_arg = 1;
@@ -436,7 +429,36 @@ handle_option_value:
         exit(0);
     }
     unsetenv("KITTY_SI_DATA");
-    if (opts.single_instance) single_instance_main(argc, argv, &opts);
+    if (opts.single_instance) {
+        char igbuf[256];
+        if (instance_group_prefix && instance_group_prefix[0]) {
+            if (opts.instance_group && opts.instance_group[0]) {
+                snprintf(igbuf, sizeof(igbuf), "%s-%s", instance_group_prefix, opts.instance_group ? opts.instance_group : "");
+                opts.instance_group = igbuf;
+            } opts.instance_group = instance_group_prefix;
+        }
+        single_instance_main(argc, argv, &opts);
+    }
+}
+
+static bool
+delegate_to_kitten_if_possible(int argc, char *argv[], char* exe_dir) {
+    if (argc > 1 && argv[1][0] == '@') exec_kitten(argc, argv, exe_dir);
+    if (argc > 2 && strcmp(argv[1], "+kitten") == 0) {
+        if (is_wrapped_kitten(argv[2])) exec_kitten(argc - 1, argv + 1, exe_dir);
+        if (strcmp(argv[2], "panel") == 0) {
+            handle_fast_commandline(argc - 2, argv + 2, "panel");
+            return true;
+        }
+    }
+    if (argc > 3 && strcmp(argv[1], "+") == 0 && strcmp(argv[2], "kitten") == 0) {
+        if (is_wrapped_kitten(argv[3])) exec_kitten(argc - 2, argv + 2, exe_dir);
+        if (strcmp(argv[3], "panel") == 0) {
+            handle_fast_commandline(argc - 3, argv + 3, "panel");
+            return true;
+        }
+    }
+    return false;
 }
 
 int main(int argc, char *argv[], char* envp[]) {
@@ -452,8 +474,7 @@ int main(int argc, char *argv[], char* envp[]) {
     if (!read_exe_path(exe, sizeof(exe))) return 1;
     strncpy(exe_dir_buf, exe, sizeof(exe_dir_buf));
     char *exe_dir = dirname(exe_dir_buf);
-    delegate_to_kitten_if_possible(argc, argv, exe_dir);
-    handle_fast_commandline(argc, argv);
+    if (!delegate_to_kitten_if_possible(argc, argv, exe_dir)) handle_fast_commandline(argc, argv, NULL);
     int num, ret=0;
     char lib[PATH_MAX+1] = {0};
     if (KITTY_LIB_PATH[0] == '/') {